Moved hardware/temperaturecontroller/__init__.py to hardware/temperaturecontroller/core.py
